+++ src/json/index.ts
@@ -1,10 +1,12 @@
-import { array, json, object, primitive, stringifiedJson } from "./json.js";
+import { array } from "./array.js";
+import { json } from "./json.js";
+import { object } from "./object.js";
+import { primitive } from "./primitive.js";
 
 const module = Object.assign(json, {
   array,
   object,
   primitive,
-  stringified: stringifiedJson,
 });
 
 export { module as json };