+++ src/string/types.ts
@@ -1,2 +1,3 @@
 export { IntegerString } from "./integer.js";
+export { JsonString } from "./json.js";
 export { UrlString } from "./url.js";